feat(relayer): update tests
