+++ packages/relayer/jest.config.js
@@ -11,6 +11,10 @@ export default {
     "^(\\.{1,2}/.*)\\.js$": "$1",
   },
   testEnvironment: "miniflare",
+  forceExit: true,
+  verbose: true,
+  detectOpenHandles: true,
+  clearMocks: true,
   testEnvironmentOptions: {
     // Miniflare doesn't yet support the `main` field in `wrangler.toml` so we
     // need to explicitly tell it where our built worker is. We also need to
