+++ packages/relayer/package.json
@@ -8,8 +8,8 @@
     "build": "node build.js",
     "dev": "miniflare --live-reload --debug --modules dist/index.mjs --env .env",
     "dev:remote": "wrangler dev",
-    "test:only": "yarn build && node --experimental-vm-modules --no-warnings node_modules/.bin/jest",
-    "test": "(cd ../contracts && yarn tests:ts) && (cd ../sdk && yarn test:only) && yarn test:only",
+    "test:only": "yarn build && node --experimental-vm-modules --no-warnings node_modules/.bin/jest --forceExit --detectOpenHandles",
+    "test": "(cd ../contracts && yarn tests:ts) && (cd ../sdk && yarn test:only) && yarn test:only --forceExit --detectOpenHandles",
     "types:check": "tsc && tsc -p test/tsconfig.json",
     "deploy": "yarn build && wrangler publish --keep-vars"
   },
@@ -18,6 +18,7 @@
   "license": "MIT",
   "devDependencies": {
     "@cloudflare/workers-types": "^3.11.0",
+    "@jest/globals": "^29.5.0",
     "@types/big.js": "^6.1.6",
     "@types/jest": "^27.5.1",
     "esbuild": "^0.14.41",
